# Conflicts:
#	Podfile.lock
#	fearless.xcodeproj/project.pbxproj
#	fearless/Common/Network/Subquery/Data/SubqueryHistory.swift

+++ fearless/Common/Model/ManagedConnectionItemMapper.swift
@@ -27,7 +27,7 @@ final class ManagedConnectionItemMapper: CoreDataMapperProtocol {
             let identifier = entity.chainId,
             let url = URL(string: identifier),
             let title = entity.name,
-            let networkType = SNAddressType(rawValue: UInt8(entity.addressPrefix))
+            let networkType = SNAddressType(rawValue: UInt16(entity.addressPrefix))
         else {
             throw ManagedAccountItemMapperError.invalidEntity
         }
